extract onnx functions into separate file
