+++ src/UI.py
@@ -8,6 +8,7 @@ import src.config as config
 import src.utils as utils
 import src.analytics as analytics
 import src.AI as AI
+import src.onnx as onnx
 
 # used to get properties of the selected style liek prompt or strangth
 # will be filled while interface is loading
@@ -51,6 +52,7 @@ def action_handle_input_file(request: gr.Request, image, token_count):
         print (e)
         gr.warning("Could not create a proper description, please describe your image shortly")
 
+    # TODO: check also skip_ai
     if config.is_feature_generation_with_token_enabled:
         #TODO: check that the image was not already used in this session
         image_sha1 = sha1(image.tobytes()).hexdigest()
@@ -66,7 +68,7 @@ def action_handle_input_file(request: gr.Request, image, token_count):
             session_image_hashes[image_sha1]=datetime.now()+timedelta(hours=4)# TODO: 4h aus konfiguration lesen
             detected_faces = []
             try:
-                detected_faces = utils.get_gender_and_age_from_image(image)
+                detected_faces = onnx.get_gender_and_age_from_image(image)
             except Exception as e:
                 print (e)
